Revert "Control Update"
This reverts commit 196c40ec21ae33b459ba58799473e406ae952c30.

+++ Assets/Scenes/Kaan/kaan script/PlayerStats.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public class PlayerStats : MonoBehaviour
 {
-    float waterSize = 0;
+    double waterSize = 0;
 
 
     public void addWater(){
-        waterSize += 0.2f;
+        waterSize+=0.2;
         Debug.Log(waterSize);
     }
